feat: Fetch Missions from SpaceX API and Store in Redux
In this commit, we have implemented the ability to fetch data from the SpaceX Missions API endpoint (https://api.spacexdata.com/v3/missions) when a user navigates to the Missions section of the application.

- We added a new action in 'actions/missions.js' to perform the API request using Axios and store the retrieved data.
- A corresponding reducer in 'reducers/missions.js' was created to handle the state updates when the data is received.
- The 'Missions' component in 'components/Missions.js' was updated to dispatch the 'fetchMissions' action when the component mounts using the 'useEffect' hook.
- The retrieved missions data is stored in the Redux store and is made available for rendering in the 'Missions' component.

This enhancement allows the Missions section to dynamically fetch and display the SpaceX mission data, providing users with the latest information when they access this section of the application.